src/requirements.txt
python-novaclient==2.20.0
python-keystoneclient==0.11.1
# 2.3.6 version not allow connect using session, only user username, passwd... 
# but using session is not allowed in some environments and this is a problem 
# with the perfomance.
#python-neutronclient==2.3.6 
python-neutronclient==2.3.9
configparser
